metrics: Add MeterProvider, 2 new flush metrics
Adds 2 new arguments to `Config`; `MeterProvider`, `MetricAttributes`.
The `MeterProvider` is used to create and record two new metrics,
`elasticsearch.buffer.latency` and `elasticsearch.flushed.latency`. The
former measures how long a document is stored in the Bulk Indexer's
buffer for, while the latter records the Elasticsearch _bulk request
latency.

`elasticsearch.buffer.latency` does not take into account the time a
document spent in the buffered channel. Doing so is quite costly and
reduces the performance more than expected, so it was abandoned.

`elasticsearch.flushed.latency` not only measures the strict request
round trip, but also the response parsing and atomic metric operations.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>