feat: add FlushTimeout config and use context timeout for flush requests
The ES client is using the err group context when flushing which means the
bulk index request could hang forever.
This PR adds a FlushTimeout option which is passed to the flush context,
ensuring a deadline is always present.
For backward compatibility and to avoid changing behaviour the default
value is no timeout.