feat: return an error when using small flush bytes with compression enabled
When compression is enabled the gzip writer will buffer data internally
which will cause the unerlying buffer to grow from 0 to n once compression
happens and compressed data are flushed.
Because of this, small flush bytes values are ignored and will never be hit.
The PR returns an error when using a small flush bytes value with compression
enabled.