Add BulkIndexerPool to improve resource sharing (#243)
Adds a new `BulkIndexerPool` to the appender to improve resource sharing
between appenders, so they are reused and not created by each appender.
The pool can be shared between multiple instances of the `Appender` to
reduce overhead, resource sharing and improve performance.

The change is implemented in a backward compatible way.

The `BulkIndexerPool` accepts four parameters:

- `guaranteed` defines the guaranteed number of `BulkIndexer`s that can
be leased by a given ID. Even when the total limit is exceeded, more
`BulkIndexer`s can still be leased until `guaranteed` is hit.
- `max` defines the maximum number of `BulkIndexer`s that can be leased
by a given ID. It ensures that no more than this number can be active
at any time for a specific ID.
- `total` defines the total number of `BulkIndexer`s that can be leased
for the entire pool. It limits the overall resource usage of the pool.

---------

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>
Co-authored-by: Andrew Wilkins <axwalk@gmail.com>